Gameplay improvements (#35)
* Closes #16 - undiscovered and resource tiles added

- Added a fog overlay that is illuminated after player discovers that
zone.

- Fog obscures the map but it is still visible, resource tiles are
only visible after discovered.

- Adding new issues to polish design regarding undiscovered map:
    + Have softer edge on tiles
    + Switch to a dungeon themed tile set
    + illuminate tiles only around player position
    + Make discovered resource tiles shine/animate + remain illuminated

* Add field to PlyaerInterface() for players observed map

* Make statying still a legal move, remove maze prints

* Closes #20 simplify use of PlayerInterface()

Only one copy of the PlayerInterface() exists, as a member of the
player. The maze enviornment accesses the interface through the
sprite.ai member.

* Closes #19 - Conform DFS to game enviornment

Game enviornment can now use a DFS player. Minor updates to the
enviornment have made the random_mouse player in need of updates,
reference issue #23 for progress on random_mouse

* Cloeses #21 - Droping mineral animation

Could not find an animated crystal/gem/gold pile 16x16 tileset, pushing
this out until a tileset is located, created, or commisioned

Illumination follows the player interface movement and is blocky,
reference issue #24 for a more polished iteration

* Closes #17 - remap direction to use up/down/left/right

Additionally caught an error where dfs had mapped E/W with inverted
offsets

* Addition changes to close #17

* Closes #24 - Polish illumination around sprite

- Sprite has a persistent disk of illuminated tiles centered at the
sprites location which tracks with is movements

- Minerals have a similar disk

* Closes #22 - implement a UI

User interface contains game statistics, and a minimap.

- Minimap encodes the entire maze at 1/4 scale
- Statistics are rendered using an included font, and update every time
the screen is drawn

See issue #28 to track progress on next iteration of the ui

* Add image and font assets

* Working version of animating maze construction

In need of some serious refactoring / redesign / documentation
improvements. Looking to:

- build the maze object from a construction_json rather then directly
- Make the animation of the maze a seperate class
- Handle key press alterations to the animation in a cleaner way
    + Call a "standard animation" function initially
    + On key press call a seperate function to handle alternate
animation style
    + Pass to these function the appropiate index for the step function
- Formalize/rethink the json structure in order to simplify it
    + Get rid of the color map and just use the colors directly

* Clean up maze construction animation code

- Add pause/resume buttons
- Handle changes to animation style in a cleaner way

* Closes #6 - Animate maze construction

Maze construction is animated on a per tile bases using a JSON styled
protocol for specifying the animation steps. Currently this is all done
within python using a dictionary structure in the form:

{ 'steps': [{'<some_color>' : <list of tuples specifying tiles> }] }

- Buttons that allow for animation speed up:
    + _s_: skip step - animates the entire step rather then each tile
    + _c_: animate the next step that is productive / clears walkable
tiles
    + _d_: combine c and s repeatidly to draw all cleared paths as quickly as
possible

- The animation can also be paused with _p_ and resumed with _r_

* Closes #29 - Speed up animation

Construct the background, ui, and mini map once as a single surface.
Frame rate has increased in testing from ~12 to ~18 at its slowest, but
there seems to be a bug with pygame that may be contributing to low
frame rate. Test on a linux box to see if the frame rate changes.

* Closes #30 - UI Improvements

- Maze construction includes a UI that displays:
    + Tiles cleared
    + Tiles visited
    + User controls

- Mini map has been reworked to only display tiles that have been
visited:
    + Starts all black
    + Updates using the visited_tiles member
    + Displays walkable tiles as white (0, 0, 0)
    + Displays resource tiles a light shade of blue (150, 255, 255)

* Updates to image assets